Second cleanup round
* get rid of as many extern / function definition in QEMU codebase.
* mostly moved cpu / gdb related code.
* still a bit of work to finish the cleanup
