+++ gdbstub/gdbstub.c
@@ -1648,30 +1648,6 @@ static void handle_query_thread_extra(GArray *params, void *user_ctx)
     gdb_put_strbuf();
 }
 
-//// --- Begin LibAFL code ---
-
-struct libafl_custom_gdb_cmd* libafl_qemu_gdb_cmds;
-
-void libafl_qemu_add_gdb_cmd(int (*callback)(void*, uint8_t*, size_t), void* data);
-void libafl_qemu_add_gdb_cmd(int (*callback)(void*, uint8_t*, size_t), void* data)
-{
-    struct libafl_custom_gdb_cmd* c = malloc(sizeof(struct libafl_custom_gdb_cmd));
-    c->callback = callback;
-    c->data = data;
-    c->next = libafl_qemu_gdb_cmds;
-    libafl_qemu_gdb_cmds = c;
-}
-
-void libafl_qemu_gdb_reply(const char* buf, size_t len);
-void libafl_qemu_gdb_reply(const char* buf, size_t len)
-{
-    g_autoptr(GString) hex_buf = g_string_new("O");
-    gdb_memtohex(hex_buf, (const uint8_t *) buf, len);
-    gdb_put_packet(hex_buf->str);
-}
-
-//// --- End LibAFL code ---
-
 static void handle_query_supported(GArray *params, void *user_ctx)
 {
     CPUClass *cc;
